# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/maritime/police-stations/{id}": {
      "get": {
        "tags": [
          "涉海档案 · 派出所"
        ],
        "summary": "查询派出所完整档案",
        "operationId": "detail_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePoliceStationView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "涉海档案 · 派出所"
        ],
        "summary": "编辑或启停派出所，需最新版本",
        "operationId": "update_8",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePoliceStation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePoliceStationView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "涉海档案 · 派出所"
        ],
        "summary": "删除未被引用的派出所，需最新版本",
        "operationId": "delete_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/maritime/ports": {
      "get": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "分页查询港口",
        "operationId": "list_6",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultPortView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "涉海档案 · 港口"
        ],
        "summary": "新建港口",
        "operationId": "create_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePort"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePortView"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/port-officers": {
      "get": {
        "tags": [
          "涉海档案 · 民警"
        ],
        "summary": "分页查询民警",
        "operationId": "list_7",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policeStationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultPortOfficerView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "涉海档案 · 民警"
        ],
        "summary": "新建民警",
        "operationId": "create_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePortOfficer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePortOfficerView"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/police-stations": {
      "get": {
        "tags": [
          "涉海档案 · 派出所"
        ],
        "summary": "分页查询派出所",
        "operationId": "list_8",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultPoliceStationView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "涉海档案 · 派出所"
        ],
        "summary": "新建派出所",
        "operationId": "create_9",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePoliceStation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePoliceStationView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/maritime/options/wharf-relations": {
      "get": {
        "tags": [
          "maritime-options-controller"
        ],
        "operationId": "wharfs",
        "parameters": [
          {
            "name": "policeStationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "responsibleOfficerId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultWharfRelationView"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/options/ports": {
      "get": {
        "tags": [
          "maritime-options-controller"
        ],
        "operationId": "ports",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultArchiveOption"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/options/port-officers": {
      "get": {
        "tags": [
          "maritime-options-controller"
        ],
        "operationId": "officers",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "policeStationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultArchiveOption"
                }
              }
            }
          }
        }
      }
    },
    "/api/maritime/options/police-stations": {
      "get": {
        "tags": [
          "maritime-options-controller"
        ],
        "operationId": "stations",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultArchiveOption"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/intelligence-topics/unit-options": {
      "get": {
        "tags": [
          "信息流转"
        ],
        "operationId": "options_1",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topicId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "receiptId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListIntelligenceUnitOption"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
      "WritePort": {
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "新建不传；编辑必填，缺失返回400，旧版本返回409",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "region": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "ENABLED|DISABLED"
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "name",
          "region",
          "status"
        ]
      },
      "ApiResponsePortView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PortView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
...
      "WritePortOfficer": {
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "新建不传；编辑必填，缺失返回400，旧版本返回409",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "ENABLED|DISABLED"
          },
          "policeStationId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[1-9][0-9]{0,18}"
          },
          "duty": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120,
            "minLength": 0
          }
        },
        "required": [
          "name",
          "policeStationId",
          "status"
        ]
      },
      "ApiResponsePortOfficerView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PortOfficerView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
...
      "WritePoliceStation": {
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "新建不传；编辑必填，缺失返回400，旧版本返回409",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "region": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "ENABLED|DISABLED"
          }
        },
        "required": [
          "name",
          "region",
          "status"
        ]
      },
      "ApiResponsePoliceStationView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PoliceStationView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PoliceStationView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fixtureKey": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "fixtureKey",
          "id",
          "location",
          "name",
          "region",
          "status",
          "updatedAt",
          "version"
        ]
      },
...
      "ApiResponsePageResultPortView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultPortView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PageResultPortView": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortView"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "ApiResponsePageResultPortOfficerView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultPortOfficerView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PageResultPortOfficerView": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortOfficerView"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "ApiResponsePageResultPoliceStationView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultPoliceStationView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
...
      "ApiResponsePageResultWharfRelationView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultWharfRelationView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
...
      "ApiResponsePageResultArchiveOption": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultArchiveOption"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
...
      "ApiResponseListIntelligenceUnitOption": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntelligenceUnitOption"
            }
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      }
    }
  }
}